feat(zendesk): separates zendesk/ticket route logic into separate func
